+++ CHANGELOG.md
@@ -6,6 +6,26 @@ Full documentation for ROCm Compute Profiler is available at [https://rocm.docs.
 
 ### Added
 
+* Add `rocpd` choice for `--format-rocprof-output` option in profile mode
+* Add `--retain-rocpd-output` option in profile mode to save large raw rocpd databases in workload directory
+
+### Changed
+
+* Add notice for change in default output format to `rocpd` in a future release
+  * This is displayed when `--format-rocprof-output rocpd` is not used in profile mode
+
+* When `--format-rocprof-output rocpd` is used, only pmc_perf.csv will be written to workload directory instead of mulitple csv files.
+
+### Resolved issues
+
+### Known issues
+
+### Removed
+
+## ROCm Compute Profiler 3.2.0 for ROCm 7.0.0
+
+### Added
+
 * Support Roofline plot on CLI (single run)
 
 * Stochastic (hardware-based) PC sampling has been enabled for AMD Instinct MI300X series and later accelerators.
@@ -66,6 +86,9 @@ Full documentation for ROCm Compute Profiler is available at [https://rocm.docs.
 
 * Add deprecation warning for database update mode.
 
+* Show description of metrics during analysis
+  * Use `--include-cols Description` to show `Description` column which is excluded by default from cli output
+
 ### Changed
 
 * Change the default rocprof version to rocprofv3, this is used when environment variable "ROCPROF" is not set
@@ -101,6 +124,7 @@ Full documentation for ROCm Compute Profiler is available at [https://rocm.docs.
 * Fixed not detecting memory clock issue when using amd-smi
 * Fixed standalone GUI crashing
 * Fixed L2 read/write/atomic bandwidths on MI350
+* Update metric names for better alignment between analysis configuration and documentation
 
 ### Known issues